# api/src/swagger/swagger.json
{
	"components": {
...
		"schemas": {
...
			"IPlatform": {
				"properties": {
					"idPlatformOfAdmin": {
						"type": "string"
					},
					"platformName": {
						"type": "string"
					},
					"platformId": {
						"type": "string"
					}
				},
				"required": [
					"idPlatformOfAdmin",
					"platformName",
					"platformId"
				],
				"type": "object"
			},
...
			"ServerType": {
				"enum": [
					"INTERNAL_SERVER",
					"EXTERNAL_SERVER"
				],
				"type": "string"
			},
			"CONNECTION_METHODS": {
				"enum": [
					"local",
					"saml",
					"oauth2",
					"openid connect"
				],
				"type": "string"
			},
			"IOpenIdAuthenticationInfo": {
				"properties": {
					"issuer": {
						"type": "string"
					},
					"authorizationUrl": {
						"type": "string"
					},
					"tokenUrl": {
						"type": "string"
					},
					"userInfoUrl": {
						"type": "string"
					},
					"clientId": {
						"type": "string"
					},
					"clientSecret": {
						"type": "string"
					},
					"callbackUrl": {
						"type": "string"
					},
					"scopes": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						]
					},
					"logoutUrl": {
						"type": "string"
					},
					"profileClassifyByPriority": {
						"type": "string"
					}
				},
				"required": [
					"issuer",
					"authorizationUrl",
					"tokenUrl",
					"userInfoUrl",
					"clientId",
					"clientSecret",
					"callbackUrl",
					"scopes"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISAMLAuthenticationInfo": {
				"properties": {
					"issuer": {
						"type": "string"
					},
					"entryPoint": {
						"type": "string"
					},
					"cert": {
						"type": "string"
					},
					"callbackUrl": {
						"type": "string"
					},
					"logoutUrl": {
						"type": "string"
					},
					"serverEntityId": {
						"type": "string"
					},
					"profileClassifyByPriority": {
						"type": "string"
					}
				},
				"required": [
					"issuer",
					"entryPoint",
					"cert",
					"callbackUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOAuthAuthenticationInfo": {
				"properties": {
					"clientId": {
						"type": "string"
					},
					"clientSecret": {
						"type": "string"
					},
					"callbackUrl": {
						"type": "string"
					},
					"grant_type": {
						"type": "string"
					},
					"endpoint": {
						"type": "string"
					},
					"logoutUrl": {
						"type": "string"
					},
					"scopes": {
						"type": "string"
					},
					"tokenUrl": {
						"type": "string"
					},
					"profileClassifyByPriority": {
						"type": "string"
					}
				},
				"required": [
					"clientId",
					"clientSecret",
					"tokenUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ILocalAuthenticationInfo": {
				"properties": {
					"code_challenge": {
						"type": "string"
					},
					"code_challenge_method": {
						"type": "string",
						"enum": [
							"S256"
						],
						"nullable": false
					}
				},
				"required": [
					"code_challenge",
					"code_challenge_method"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ILoginServer": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ServerType"
					},
					"authentication_method": {
						"$ref": "#/components/schemas/CONNECTION_METHODS"
					},
					"authentication_info": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/IOpenIdAuthenticationInfo"
							},
							{
								"$ref": "#/components/schemas/ISAMLAuthenticationInfo"
							},
							{
								"$ref": "#/components/schemas/IOAuthAuthenticationInfo"
							},
							{
								"$ref": "#/components/schemas/ILocalAuthenticationInfo"
							}
						]
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"INotification": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"actor": {
						"properties": {
							"actorName": {
								"type": "string"
							},
							"actorId": {
								"type": "string"
							}
						},
						"required": [
							"actorName",
							"actorId"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"name",
					"date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"INotificationCreationParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"actor": {
						"properties": {
							"actorName": {
								"type": "string"
							},
							"actorId": {
								"type": "string"
							}
						},
						"required": [
							"actorName",
							"actorId"
						],
						"type": "object"
					}
				},
				"required": [
					"name",
					"date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"statusOrgan": {
				"enum": [
					"online",
					"fail",
					"stop"
				],
				"type": "string"
			},
			"IOrgan": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"statusOrgan": {
						"$ref": "#/components/schemas/statusOrgan"
					},
					"organType": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"platformId": {
						"type": "string"
					}
				},
				"required": [
					"statusOrgan",
					"organType",
					"platformId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOrganCreationParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"organType": {
						"type": "string"
					},
					"statusOrgan": {
						"type": "string"
					},
					"platformId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"organType",
					"statusOrgan",
					"platformId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IOrganUpdateParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"organType": {
						"type": "string"
					},
					"statusOrgan": {
						"$ref": "#/components/schemas/statusOrgan"
					},
					"platformId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"organType",
					"statusOrgan",
					"platformId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IPlatformUpdateParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"redirectUrl": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"loginServerIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"grant_types": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Spinal Organ Auth API",
		"version": "1.0.0",
		"description": "API for managing authentication and authorization in the Spinal Organ system",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Spinalcom",
			"email": "developers@spinalcom.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/loginserver/create": {
			"post": {
				"operationId": "CreateLoginServer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"LoginServer"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:write"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ILoginServer"
							}
						}
					}
				}
			}
		},
		"/loginserver/all_servers": {
			"get": {
				"operationId": "GetAllLoginServer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {},
											"type": "array"
										},
										{
											"properties": {
												"error": {}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"LoginServer"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read"
						]
					}
				],
				"parameters": []
			}
		},
		"/loginserver/{serverId}": {
			"get": {
				"operationId": "GetLoginServer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"LoginServer"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "serverId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/loginserver/update/{serverId}": {
			"put": {
				"operationId": "EditLoginServer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"LoginServer"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:write"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "serverId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ILoginServer"
							}
						}
					}
				}
			}
		},
		"/loginserver/delete/{serverId}": {
			"delete": {
				"operationId": "DeleteLoginServer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {},
												"error": {}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"LoginServer"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "serverId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/logs": {
			"get": {
				"operationId": "GetLogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {},
											"type": "array"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Logs"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read"
						]
					}
				],
				"parameters": []
			}
		},
		"/logs/getPlatformsLogs": {
			"get": {
				"operationId": "GetPlatformsLogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {},
											"type": "array"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Logs"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read"
						]
					}
				],
				"parameters": []
			}
		},
		"/notification": {
			"post": {
				"operationId": "CreateNotification",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/INotification"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Notification"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/INotificationCreationParams"
							}
						}
					}
				}
			}
		},
		"/organs/{platformId}": {
			"post": {
				"operationId": "CreateOrgan",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IOrgan"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Organs"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:write"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IOrganCreationParams"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetOrgans",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IOrgan"
											},
											"type": "array"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Organs"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "platformId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/organs/{organId}": {
			"put": {
				"operationId": "UpdatePlateform",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IOrgan"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Organs"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "organId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IOrganUpdateParams"
							}
						}
					}
				}
			}
		},
...
		"/platforms/{platformId}": {
			"get": {
				"operationId": "GetPlateform",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IPlatform"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Platforms"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "platformId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeletePlatform",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										},
										"message": {
											"type": "string"
										}
									},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Platforms"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "platformId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePlateform",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IPlatform"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Platforms"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:write"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "platformId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPlatformUpdateParams"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}